chore: remove yarn.lock and switch to npm
After evaluating the pros and cons of continuing with Yarn, we've decided to switch to npm for package management. This decision is influenced by recent discussions and insights shared by the community:

- Insightful article by Jared Wilcurt on the Bun hype and learnings from Yarn: https://dev.to/thejaredwilcurt/bun-hype-how-we-learned-nothing-from-yarn-2n3j
- Twitter discussion by Stephen Fluin highlighting considerations for package managers: https://x.com/stephenfluin/status/1637428030531313664

By switching to npm, we aim to streamline our dependency management and leverage npm's native features and updates.